Continue to refactor the options utils
