+++ lib/lev/utils.rb
@@ -2,9 +2,8 @@ require 'lev/utils/symbolify'
 require 'lev/utils/nameify'
 require 'lev/utils/aliasify'
 
-require 'lev/utils/outputs'
-require 'lev/utils/subroutines'
 require 'lev/utils/options'
+require 'lev/utils/subroutines'
 
 module Lev
   module Utils
