+++ lib/lev/utils/options.rb
@@ -3,8 +3,7 @@ module Lev
     class Options
       def self.setup(routine_class, options)
         setup_routine_getters(routine_class, options)
-        Outputs.setup(routine_class, options.delete(:outputs))
-        Subroutines.setup(routine_class, options.delete(:uses))
+        Subroutines.setup(routine_class, options)
       end
 
       private
